Add files via upload
This is a game of snakes and ladders but with added rules. the game follows usual snakes and ladders rules but also new ones, such as: if  a player gets a six on their dice , they will be asked to choose from two options, both benefitting them. also if a player lands on certain numbers , they will be given random mathematical problems to solve in a given time limit, if they do solve it in time they will be rewarded or else they will be moved back. However the hardest barrier of the game is at the 100th box. When a player reaches box 100, they will have to play a hangman game and if they solve it successfully, they win the game or else they will be moved back. the hangman game gets it's words from the .txt file and i have based it on sports but, any word can be added to the file to increase the word box of the game.